default validator + move common logic to extensible validator
