+++ FluentValidation/Library/Validators/FriendlyPropertyValidator.cs
@@ -1,39 +1,42 @@
-﻿using FluentValidation.Validators;
-using Woody230.FluentValidation.Message;
+﻿using FluentValidation;
+using FluentValidation.Validators;
 
 namespace Woody230.FluentValidation.Validators
 {
     /// <summary>
+    /// <para>
     /// Validates a <typeparamref name="TProperty"/> within the context of a <typeparamref name="T"/>.
+    /// </para>
+    /// <para>
+    /// When the property is null, then the property is considered valid.
+    /// This allows the reuse of logic within an optional/required context where a <see cref="NotNullValidator{T, TProperty}"/> is used if the value should be considered required.
+    /// </para>
     /// </summary>
-    public abstract class FriendlyPropertyValidator<T, TProperty> : PropertyValidator<T, TProperty>
+    public abstract class FriendlyPropertyValidator<T, TProperty> : ExtensiblePropertyValidator<T, TProperty>
     {
-        /// <inheritdoc/>
-        public override string Name
+        /// <summary>
+        /// Validates whether the <paramref name="value"/> is valid.
+        /// Null values are automatically considered valid and should be handled separately with a <see cref="NotNullValidator{T, TProperty}"/> if the value is required.
+        /// </summary>
+        /// <param name="context">The context.</param>
+        /// <param name="value">The nullable value.</param>
+        /// <returns>True if the value is null or valid, otherwise false.</returns>
+        public override bool IsValid(ValidationContext<T> context, TProperty? value)
         {
-            get
+            if (value == null)
             {
-                var name = GetType().Name;
-                return name[..name.IndexOf('`')];
+                return true;
             }
+
+            return IsPropertyValid(context, value);
         }
 
         /// <summary>
-        /// Gets the default error message for this validator.
-        /// The message includes the name of the property at the beginning.
+        /// Validates whether the <paramref name="value"/> is valid.
         /// </summary>
-        /// <param name="errorCode">The error code.</param>
-        /// <returns>The default error message.</returns>
-        protected override string GetDefaultMessageTemplate(string errorCode)
-        {
-            return BuildDefaultErrorMessage(errorCode).Prepend("'").PrependPropertyName().Prepend("'").ToString();
-        }
-
-        /// <summary>
-        /// Builds the default error message for this validator.
-        /// </summary>
-        /// <param name="errorCode">The error code.</param>
-        /// <returns>The error message builder.</returns>
-        protected abstract ErrorMessageTemplateBuilder BuildDefaultErrorMessage(string errorCode);
+        /// <param name="context">The context.</param>
+        /// <param name="value">The value.</param>
+        /// <returns>True if valid, otherwise false.</returns>
+        protected abstract bool IsPropertyValid(ValidationContext<T> context, TProperty value);
     }
 }
